requirements.txt
pygame==2.6.1
numpy==2.1.3
matplotlib==3.9.2
requests==2.32.3
openai==1.69.0
python-dotenv==1.0.1